Create Solstice_Commands.py
This is a python program to get info from a solstice pod on your network and to send basic commands to the pod.

When sending commands it will open the URL in a new tab of the Windows default browser. It will then close the tab after 5 seconds to ensure the command was passed. Some commands stay on an infinite loop if you do not close the tab.